Cluster events do not care about the domain object any more
